+++ src/app/management/components/header/management-header.component.ts
@@ -13,8 +13,8 @@ import { Observable } from 'rxjs';
 import { pluck } from 'rxjs/operators';
 import { AppService } from 'src/app/app.service';
 import { ManagementService } from 'src/app/management/management.service';
-import { ConfirmationDialogComponent } from 'src/app/shared/dialogs/confirmation-dialog/confirmation-dialog.component';
-import { ConfirmationDialogData } from 'src/app/shared/dialogs/confirmation-dialog/ConfirmationDialogData';
+import { ConfirmationDialogComponent } from 'src/app/shared/dialogs/confirmation/confirmation-dialog.component';
+import { ConfirmationDialogData } from 'src/app/shared/dialogs/confirmation/ConfirmationDialogData';
 import { EditProfileFormDialogComponent } from 'src/app/shared/dialogs/edit-profile-form-dialog/edit-profile-form-dialog.component';
 import { environment } from 'src/environments/environment';
 import { COMMON_DISMISS_BUTTON_LABEL } from 'src/text/messages';
