+++ src/app/shared/dialogs/confirmation/ConfirmationDialogData.ts
