+++ src/app/shared/shared.module.ts
@@ -30,7 +30,7 @@ import { ShipperFormComponent } from './components/shipper-form/shipper-form.com
 import { SlideshowComponent } from './components/slideshow/slideshow.component';
 import { UserFormComponent } from './components/user-form/user-form.component';
 import { AddressFormDialogComponent } from './dialogs/address-form/address-form-dialog.component';
-import { ConfirmationDialogComponent } from './dialogs/confirmation-dialog/confirmation-dialog.component';
+import { ConfirmationDialogComponent } from './dialogs/confirmation/confirmation-dialog.component';
 import { EntityFormDialogComponent } from './dialogs/entity-form-dialog/entity-form-dialog.component';
 import { EditProfileFormDialogComponent } from './dialogs/edit-profile-form-dialog/edit-profile-form-dialog.component';
 import { ImagesArrayDialogComponent } from './dialogs/images-array/images-array-dialog.component';
